requirements.txt
Flask==2.3.3
reportlab==4.0.4
Werkzeug==2.3.7
requests==2.31.0
python-dotenv==1.0.0
